requirements.txt
# AEMplify - Streamlit UI Requirements

# Core Framework
streamlit>=1.31.0

# CrewAI and Dependencies
crewai>=0.70.0
crewai-tools>=0.12.0

# AI/LLM
anthropic>=0.18.0
openai>=1.12.0

# File handling and utilities
Pillow>=10.0.0
python-dotenv>=1.0.0

# Web and image processing
requests>=2.31.0
